contributors/i18n/submitted/css3-syntax/i18n-css-syntax-bytes-018.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="UTF-8"/>
<title>CSS3 Syntax, input byte stream: css bom and @charset iso15</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-syntax/#input-byte-stream'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='http dom'>
<meta name="assert" content="When a stylesheet has a BOM and a @charset declaration that is not for a Unicode encoding, the BOM takes precedence.">
<style type='text/css'>
.test div { width:50px; }

p { color: gray; margin-top: 3em; font-size: 80%; }
</style>
<link rel="stylesheet" type="text/css" href="support/bom-charset15.css">
</head>
<body>



<div class="test"><div id="box" class="ýäè">&#xA0;</div></div>


<!--Notes:
This test is only valid if the user agent recognises the @charset and BOM declarations.

In the style sheet, the CSS selector identifies a class name with the following byte sequence: c3 bd c3 a4 c3 a8.  Read as UTF-8 this gives the characters ýäè.  Read as ISO 8859-15, this gives ÃœÃ€Ãš.

This HTML page is served as UTF-8 and uses the same byte sequence for the class name that equates to the character sequence ýäè.  

The class ids will therefore only match if the CSS is read as UTF-8.

The CSS starts with a UTF-8 signature but also has an @charset rule declaring the encoding to be ISO 8859-15. There is no HTTP encoding information. The browser should set the encoding to UTF-8, because of the BOM, and ignore the @charset rule.

The class value above will only match if the CSS is read by the browser as UTF-8.

-->
<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, " ");
</script>

<div id='log'></div>

<p class="ýäè"><span><span id="debug">Text as read from the style sheet: </span><br />(If blank, the style sheet hasn't been read at all.)</span></p>
</body>
</html>
